feat: separated auth module files from usr module files
